Version 4.0
This version is tweaked a bit. I muted the colors a little and also eliminated some of the lower spectrum highlighting in favor of more blues and greens. This seems to be easier to look at, in my opinion. I also added some custom highlights for the vimwiki script (vimscript #2226).