Version 6.1.0
Switched around some of the main highlight colors. Instead of using a green for constants, I switched it to be the yellow/orange. This makes constants more distinguishable from other highlights (identifier, statement, preproc, and type are now all some sort of blue or green). Also tweaked a few other settings. The "blue" style is kind of back, but is actually just the dark style with a more blue background.