Version 7.0.0
Version 7 has unified the 256 color terminal and GUI versions (the GUI version only uses colors available on the 256 color terminal). The overall colors were also toned down a little bit (light version is now a light gray instead of white and the dark version is slightly lighter) to make it easier on the eyes.